+++ packages/snap-client/src/Client/utils/fibonacci.ts
@@ -1,0 +1,14 @@
+export const fibonacci = (num: number): number => {
+	let a = 1,
+		b = 0,
+		temp;
+
+	while (num >= 0) {
+		temp = a;
+		a = a + b;
+		b = temp;
+		num--;
+	}
+
+	return b;
+};